requirements.txt
speedtest-cli
rich

